feat: create seeder

+++ database/migrations/2014_02_08_171603_create_positions_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('division_id');
             $table->string('name');
-            $table->float('salary_per_hour');
+            $table->bigInteger('salary_per_hour');
             $table->timestamps();
 
             $table->foreign('division_id')->references('id')->on('divisions')->onDelete('cascade');
